function for drive takes fuel from engine in class added, code errors unresolved

+++ src/main/java/vehicles/Refuelable.java
@@ -1,5 +1,5 @@
 package vehicles;
 
 public interface Refuelable {
-    double refuel(int distance);
+    double refuel(double distance);
 }
